requirements.txt
absl-py==0.8.0
astor==0.8.0
beautifulsoup4==4.8.2
cachetools==3.1.1
certifi==2019.11.28
cffi==1.13.1
chardet==3.0.4
cloudpickle==1.2.2
cluster==1.4.1.post2
cycler==0.10.0
Cython==0.29.14
cytoolz==0.10.0
dask==2.6.0
decorator==4.4.1
gast==0.2.2
google-auth==1.6.3
google-auth-oauthlib==0.4.1
google-pasta==0.1.7
grpcio==1.24.3
h5py==2.9.0
idna==2.8
imageio==2.6.1
joblib==0.13.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
labelbox==2.4
libtiff==0.4.2
Markdown==3.1.1
matplotlib==3.1.1
mkl-fft==1.0.14
mkl-random==1.1.0
mkl-service==2.3.0
networkx==2.4
numpy==1.17.2
oauthlib==3.1.0
olefile==0.46
opencv-python==4.1.2.30
openslide-python==1.1.1
opt-einsum==3.1.0
pandas==0.25.2
patsy==0.5.1
Pillow==6.2.0
protobuf==3.9.2
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycparser==2.19
pyparsing==2.4.2
python-dateutil==2.8.0
python-resize-image==1.1.19
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.3
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
scikit-image==0.15.0
scikit-learn==0.21.3
scipy==1.3.1
seaborn==0.9.0
segmentation==0.1.3
selenium==3.141.0
semisup-learn==0.0.1
six==1.12.0
sklearn==0.0
soupsieve==1.9.5
statsmodels==0.10.1
tb-nightly==2.1.0a20191101
tensorboard==2.0.0
tensorflow==2.0.0
tensorflow-estimator==2.0.0
termcolor==1.1.0
toolz==0.10.0
torch==1.3.0
torchvision==0.4.1a0+d94043a
tornado==6.0.3
tqdm==4.41.1
urllib3==1.25.6
webencodings==0.5.1
Werkzeug==0.16.0
wrapt==1.11.2
